requirements.txt
python-telegram-bot==20.7a0
speechrecognition
requests
gTTS
pydub